2011-12-05	Ralf Corsépius <ralf.corsepius@rtems.org>
	* libcsupport/src/termios.c: Remove unnecessary cast.

+++ cpukit/libcsupport/src/termios.c
@@ -1002,7 +1002,7 @@ fillBufferQueue (struct rtems_termios_tty *tty)
 {
   rtems_interval timeout = tty->rawInBufSemaphoreFirstTimeout;
   rtems_status_code sc;
-  int               wait = (int)1;
+  int               wait = 1;
 
   while ( wait ) {
     /*
